refactor(backend): isolate static and special transport boundaries
